+++ design/src/main/java/com/urlaunched/android/design/ui/otptextfield/constants/OtpTextFieldDimens.kt
@@ -1,8 +1,8 @@
-package com.urlaunched.android.design.ui.otpContainer.constants
+package com.urlaunched.android.design.ui.otptextfield.constants
 
 import androidx.compose.ui.unit.dp
 
-internal object OtpContainerDimens {
+internal object OtpTextFieldDimens {
     val defaultOtpCellHeight = 48.dp
     val defaultOtpCellWidth = 40.dp
     val defaultBorderWidth = 1.dp
